Refactor
Followed video tutorial for refactoring
