+++ index.html
@@ -2,7 +2,11 @@
 <html>
   <head>
     <meta charset="utf-8">
+    <meta lang="en-us">
+    <title>GhostBusters</title>
   </head>
   <body>
+    <script src="phaser.js"></script>
+    <script src="game.js"></script>
   </body>
 </html>
