+++ src/config/config.js
@@ -1,0 +1,6 @@
+export default {
+    type: Phaser.AUTO,
+    parent: "phaser-example",
+    width: 800,
+    height: 600,
+  };
